tested for floating array

+++ bubble-sort/bubblesort.h
@@ -1,3 +1,3 @@
-typedef int (*compare)(void* first, void* second) ;
+typedef float (*compare)(void* first, void* second) ;
 
 void bubbleSort(void* base, int numberOfElements, int elementSize , compare comp);
